resolved favlist update issue
The wishlist button wasn't updating correctly. When you two or more items to wishlist, and if your try to remove items from the wishlist from the wishlist section, the item below the "to be removed item" also shows option of adding it to wishlist, even though it already is in it. Possible cause are because of how react keeps track of listed elements.